ErrorHandler improved in authenticate_request

+++ app/controllers/application_controller.rb
@@ -6,16 +6,21 @@ class ApplicationController < ActionController::API
   private
 
   def authenticate_request
-    decrypted_data = UserToken.decode(token_from_header)
+    token = token_from_header
+    if token.nil?
+      raise UnauthorizedError.new('Token is missing or invalid')
+    end
+
+    decrypted_data = UserToken.decode(token)
     user_id = decrypted_data[:user_id] if decrypted_data
-    user_token = UserToken.find_by(user_id: user_id, token: token_from_header)
+    user_token = UserToken.find_by(user_id: user_id, token: token)
 
     if user_token.nil?
-      render json: { error: 'Unauthorized' }, status: :unauthorized
+      raise UnauthorizedError.new('Invalid token')
     elsif user_token.expires_at < Time.current
       user_token.destroy
       user_token.user.session.destroy
-      render json: { error: 'Token expired' }, status: :unauthorized
+      raise UnauthorizedError.new('Token expired')
     else
       @current_user = user_token.user
     end
